add some sanity check in destructor of BufferedColumnTable

+++ src/main/cpp/query_table/buffered_column_table.h
@@ -349,7 +349,11 @@ namespace vaultdb {
         ~BufferedColumnTable() {
             this->conf_.bpm_.removeTable(this->table_id_);
             if (std::filesystem::exists(this->file_path_)) {
-                std::filesystem::remove_all(this->file_path_);
+                std::filesystem::remove_all(this->secret_shares_path_);
+
+                if(std::filesystem::is_empty(this->file_path_)) {
+                    std::filesystem::remove_all(this->file_path_);
+                }
             }
         }
     };
